fix: resolve failing unit tests after code reorganization
Fix import paths and test configurations that were broken after the
code reorganization:
- Fix import paths in chatbot tests to use common.validation_utils
- Fix import paths in similarity tests to use similarity.similarity_utils
- Add conftest.py for summarize service tests
- Consolidate test dependencies in services/requirements-test.txt
- Remove duplicate test file (services/similarity/test_similarity_search.py)

Additionally, complete the refactoring from python-base to service-base:
- Rename images/python-base to images/service-base for clarity
- Remove deprecated workflow files (spyre-rag-*.yml)
- Add new workflow files for service-base, chatbot-ui, and service-unit-tests
- Update documentation to reflect new naming convention
- Update Containerfile references from python-base to service-base

This ensures all unit tests pass and aligns naming with the actual
purpose of the base image as a foundation for all services.
