+++ .github/workflows/chatbot-ui.yml
@@ -1,4 +1,4 @@
-name: Spyre RAG UI
+name: Chatbot UI
 on:
   push:
     paths:
@@ -8,7 +8,7 @@ on:
   pull_request:
     paths:
       - 'ui/chatbot/**'
-      - .github/workflows/spyre-rag-ui-image.yml
+      - .github/workflows/chatbot-ui.yml
     branches:
       - main
 
@@ -72,12 +72,12 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v5
       
-      - name: Build spyre-rag-ui image
+      - name: Build chatbot-ui image
         working-directory: ui/chatbot
         run: |
           make build REGISTRY=${{ env.ICR_REGISTRY }}/${{ env.ICR_NAMESPACE }}
 
-      - name: Publish spyre-rag-ui image
+      - name: Publish chatbot-ui image
         working-directory: ui/chatbot
         if: github.event_name == 'push' && github.ref == 'refs/heads/main'
         run: |
