--- .github/workflows/spyre-rag-base.yml
@@ -1,43 +1,0 @@
-name: Python Base
-on:
-  push:
-    paths:
-      - 'images/python-base/**'
-    branches:
-      - main
-  pull_request:
-    paths:
-      - 'images/python-base/**'
-      - .github/workflows/spyre-rag-base.yml
-    branches:
-      - main
-
-env:
-  ICR_REGISTRY: icr.io
-  ICR_NAMESPACE: ai-services-cicd
-
-jobs:
-  build-publish-image:
-    name: Build and Publish Image
-    runs-on: ubuntu-24.04-ppc64le-p10
-
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v5
-
-      - name: Build python-base image
-        working-directory: images/python-base
-        run: |
-          make build REGISTRY=${{ env.ICR_REGISTRY }}/${{ env.ICR_NAMESPACE }}
-
-      - name: Publish python-base image
-        working-directory: images/python-base
-        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-        run: |
-          make push REGISTRY=${{ env.ICR_REGISTRY }}/${{ env.ICR_NAMESPACE }} REGISTRY_USER=${{ secrets.ICR_USERNAME }} REGISTRY_PASSWORD=${{ secrets.ICR_APIKEY }}
-
-  scan:
-    uses: ./.github/workflows/scanner.yml
-    with:
-      path: 'images/python-base'
-      image: 'python-base'
