--- .github/workflows/spyre-rag-unit-tests.yml
@@ -1,52 +1,0 @@
-name: Services Unit Tests
-
-on:
-  pull_request:
-    branches:
-      - main
-    paths:
-      - 'services/**'
-      - '.github/workflows/spyre-rag-unit-tests.yml'
-
-jobs:
-  test:
-    runs-on: ubuntu-24.04-ppc64le
-    
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v6
-
-      - name: Set up Python 3.12
-        uses: IBM/setup-python-pz@291287e36d7c893a50c857d8bc014d67edd58ada
-        with:
-          python-version: '3.12'
-
-      - name: Install base dependencies
-        run: |
-          pip install -r images/python-base/requirements.txt \
-          --extra-index-url https://wheels.developerfirst.ibm.com/ppc64le/linux \
-          --prefer-binary
-      
-      - name: Install digitize test dependencies
-        run: |
-          pip install -r digitize/tests/requirements-test.txt \
-          --extra-index-url https://wheels.developerfirst.ibm.com/ppc64le/linux \
-          --prefer-binary
-        working-directory: services
-      
-      - name: Run digitize service tests
-        run: |
-          pytest digitize/tests/ -v
-        working-directory: services
-      
-      - name: Install chatbot test dependencies
-        run: |
-          pip install -r chatbot/tests/requirements-test.txt \
-          --extra-index-url https://wheels.developerfirst.ibm.com/ppc64le/linux \
-          --prefer-binary
-        working-directory: services
-      
-      - name: Run chatbot service tests
-        run: |
-          pytest chatbot/tests/ -v
-        working-directory: services
