+++ docs/proposals/code_reorg_proposal.md
@@ -88,7 +88,7 @@ project-ai-services/
 │       └── src/
 │
 ├── images/                        # Base images and utilities
-│   ├── python-base/              # Renamed from rag-base
+│   ├── service-base/             # Renamed from rag-base
 │   │   ├── Containerfile
 │   │   ├── requirements.txt
 │   │   ├── download_docling_models.py
@@ -146,7 +146,7 @@ project-ai-services/
 **Purpose:** Foundation images that services build upon.
 
 **Key Changes:**
-- **`images/python-base/`**: Renamed from `images/rag-base/` to reflect its general-purpose nature
+- **`images/service-base/`**: Renamed from `images/rag-base/` to reflect its general-purpose nature
 - Contains shared Python dependencies, docling models, system packages
 
 **Why this structure:**
@@ -183,7 +183,7 @@ catalog-ui/*                     → ui/catalog/*
 
 ### In `images/`:
 ```
-images/rag-base/*                → images/python-base/*
+images/rag-base/*                → images/service-base/*
 ```
 
 ### Delete entire directory:
@@ -193,9 +193,9 @@ spyre-rag/                       → DELETE (contents redistributed)
 
 ## Container Build Strategy
 
-### Base Image Layer (`images/python-base/`)
+### Base Image Layer (`images/service-base/`)
 ```dockerfile
-# images/python-base/Containerfile
+# images/service-base/Containerfile
 FROM registry.access.redhat.com/ubi9/ubi:9.7
 
 # Install system dependencies
@@ -209,7 +209,7 @@ RUN pip install -r requirements.txt
 COPY download_docling_models.py .
 RUN python download_docling_models.py
 
-# This image is tagged as: icr.io/ai-services-cicd/python-base:latest
+# This image is tagged as: icr.io/ai-services-cicd/service-base:latest
 ```
 
 **Contains:**
@@ -221,7 +221,7 @@ RUN python download_docling_models.py
 ### Common Library Layer (`services/common/`)
 ```dockerfile
 # services/common/Containerfile
-FROM icr.io/ai-services-cicd/python-base:latest
+FROM icr.io/ai-services-cicd/service-base:latest
 
 WORKDIR /opt/services/common
 
@@ -266,7 +266,7 @@ CMD ["/var/venv/bin/python", "-m", "uvicorn", "app:app", "--host", "0.0.0.0", "-
 ### Build Order
 ```bash
 # 1. Build base image with system deps and common Python packages
-cd images/python-base && podman build -t icr.io/ai-services-cicd/python-base:latest -f Containerfile .
+cd images/service-base && podman build -t icr.io/ai-services-cicd/service-base:latest -f Containerfile .
 
 # 2. Build common library layer
 cd services/common && podman build -t icr.io/ai-services-cicd/services-common:latest -f Containerfile .
