+++ images/service-base/docling/0001-Exclude-source-that-has-commercial-license_v0.22.1.patch
