+++ images/service-base/docling/Containerfile
