+++ images/service-base/docling/docling_v2.60.1.patch
