+++ images/service-base/download_docling_models.py
