+++ images/service-base/requirements.txt
