+++ services/common/requirements.txt
@@ -1,2 +1,2 @@
-# common-specific dependencies beyond python-base
-# All heavy deps (fastapi, opensearch-py, docling, etc.) live in images/python-base/requirements.txt
+# common-specific dependencies beyond service-base
+# All heavy deps (fastapi, opensearch-py, docling, etc.) live in images/service-base/requirements.txt
