--- services/summarize/tests/requirements-test.txt
@@ -1,21 +1,0 @@
-# Testing Dependencies for summarize service tests
-# Install with: pip install -r tests/requirements-test.txt
-
-# Core testing framework
-pytest>=7.4.0
-pytest-asyncio>=0.21.0
-pytest-cov>=4.1.0
-pytest-mock>=3.11.0
-
-# FastAPI testing
-httpx>=0.24.0
-fastapi>=0.104.0
-starlette>=0.27.0
-uvicorn==0.38.0
-requests==2.33.0
-python-multipart==0.0.27
-
-# Additional utilities
-python-dotenv>=1.0.0
-pydantic>=2.0.0
-pydantic-settings==2.13.1
